# trustbite-docs/04_Software_Engineering/openapi.yaml
openapi: 3.1.0
info:
  title: TrustBite MVP API
  version: 1.1.1
  description: Baseline OpenAPI contract for TrustBite MVP P0. Markdown specs remain the narrative source; this file is the
    machine-readable contract for development, mocks and contract tests.
servers:
- url: /api/v1
security:
- bearerAuth: []
...
paths:
...
  /users/me:
    get:
      tags:
      - Users
      summary: Get current user
      operationId: getMe
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    patch:
      tags:
      - Users
      summary: Update current user profile
      description: avatarUrl, when supplied, must reference a TrustBite-owned allowlisted media origin or a URL created by the platform avatar upload flow; arbitrary external URLs must be rejected.
      operationId: updateMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  maxLength: 120
                dateOfBirth:
                  type: string
                  format: date
                phoneNumber:
                  type: string
                  description: E.164 or a Vietnamese ten-digit local number.
                avatarUrl:
                  type: string
                  format: uri
                  description: Must be a TrustBite-owned allowlisted media URL or a URL issued by the platform avatar upload flow; arbitrary external URLs are not accepted.
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
  /users/me/avatar-upload-url:
    post:
      tags:
      - Users
      summary: Create avatar upload URL
      description: Requires authentication. Returns a short-lived upload URL for a TrustBite-owned media object and binds the requested fileSizeBytes as the signed S3 Content-Length; avatarUrl on PATCH /users/me must reference the returned public media URL after upload completion. For custom CDN or path-style S3 hosts, the returned avatarUrl path includes the bucket before avatars/... so account deletion can clean up the owned object.
      operationId: createAvatarUploadUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - contentType
              - fileSizeBytes
              properties:
                contentType:
                  type: string
                  enum:
                  - image/jpeg
                  - image/png
                  - image/webp
                fileSizeBytes:
                  type: integer
                  minimum: 1
                  maximum: 5242880
      responses:
        '200':
          description: Avatar upload URL created
          content:
            application/json:
              schema:
                type: object
                required:
                - uploadUrl
                - avatarUrl
                - expiresAt
                properties:
                  uploadUrl:
                    type: string
                    format: uri
                    description: Short-lived signed URL for uploading the avatar image.
                  avatarUrl:
                    type: string
                    format: uri
                    description: TrustBite-owned allowlisted media URL accepted by PATCH /users/me after upload; custom CDN and path-style S3 hosts include /<bucket>/avatars/... for deletion cleanup compatibility.
                  expiresAt:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
...
  /gamification/me:
    get:
      tags:
      - Gamification
      summary: Get current user gamification summary
      operationId: getMyGamification
      responses:
        '200':
          description: Gamification summary
          content:
            application/json:
              schema:
                type: object
                required:
                - expPoints
                - rankCode
                - badges
                properties:
                  expPoints:
                    type: integer
                  rankCode:
                    type: string
                  badges:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly Secure SameSite=Strict refresh token cookie issued by auth endpoints; raw token values must not be exposed to JavaScript or stored in localStorage.
    adminBffSecret:
      type: apiKey
      in: header
      name: x-trustbite-bff-secret
      description: Server-only credential used only between Next.js and Express.
    adminWebSession:
      type: apiKey
      in: header
      name: x-trustbite-admin-session
      description: Opaque admin session marker forwarded by the Next.js BFF; never exposed to browser JavaScript.
...
  responses:
    Error:
      description: Standard error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    User:
      type: object
      required:
      - id
      - rankCode
      - status
      - expPoints
      - profileComplete
      properties:
        id:
          type: string
          format: uuid
        phoneNumber:
          type: string
          nullable: true
          description: Mask in admin/public contexts.
        displayName:
          type: [string, 'null']
          nullable: true
        dateOfBirth:
          type: string
          format: date
          nullable: true
        profileComplete:
          type: boolean
        role:
          type: string
          enum:
          - USER
          - MERCHANT
          - ADMIN
          - SUPER_ADMIN
        roles:
          type: array
          description: TrustBite-local roles used for server-side authorization.
          items:
            type: string
            enum:
            - USER
            - MERCHANT
            - ADMIN
            - SUPER_ADMIN
        status:
          type: string
          enum:
          - ACTIVE
          - REVIEW_RESTRICTED
          - SUSPENDED
          - DELETED
        expPoints:
          type: integer
        rankCode:
          type: string
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              enum:
              - VALIDATION_ERROR
              - AUTH_REQUIRED
              - TOKEN_EXPIRED
              - FORBIDDEN
              - NOT_FOUND
              - RATE_LIMITED
              - OTP_INVALID
              - OTP_EXPIRED
              - FILE_TOO_LARGE
              - UNSUPPORTED_FILE_TYPE
              - DUPLICATE_RECEIPT_HASH
              - REVIEW_NOT_EDITABLE
              - RESTAURANT_NOT_ACTIVE
              - RESTAURANT_NOT_FOUND
              - CLAIM_CONFLICT
              - REPORT_DUPLICATE
              - PROVIDER_UNAVAILABLE
              - ACCOUNT_SUSPENDED
              - ACCOUNT_DELETED
              - IDEMPOTENCY_KEY_REQUIRED
              - IDEMPOTENCY_KEY_INVALID
              - IDEMPOTENCY_CONFLICT
              - REQUEST_IN_PROGRESS
              - REVIEW_VERIFICATION_EXPIRED
              - ADMIN_REASON_REQUIRED
              - DELETION_REQUEST_ALREADY_EXISTS
              - DELETION_REQUEST_NOT_FOUND
              - CANNOT_BLOCK_SELF
              - USER_ALREADY_BLOCKED
              - GPS_PERMISSION_REQUIRED
              - INVALID_GPS_SIGNAL
              - INVALID_RECEIPT_DECISION_COMBINATION
              - INVALID_MODERATION_DECISION_COMBINATION
              - REVIEW_NOT_FOUND
              - REVIEW_NOT_VISIBLE
              - UNSUPPORTED_TARGET_LOCALE
              - TRANSLATION_TEXT_EMPTY
              - TRANSLATION_RATE_LIMITED
              - TRANSLATION_PROVIDER_UNAVAILABLE
              - USER_ALREADY_SUSPENDED
              - USER_NOT_SUSPENDED
              - CANNOT_SUSPEND_SELF
              - CANNOT_SUSPEND_DELETED_USER
              - CANNOT_REACTIVATE_SELF
              - CANNOT_REACTIVATE_DELETED_USER
              - ADMIN_SESSION_INVALID
              - ADMIN_SESSION_EXPIRED
              - ADMIN_ACCESS_REQUIRED
              - DELETION_REQUEST_ACTIVE
              - USER_NOT_FOUND
              - EMAIL_IN_USE
              - PHONE_NUMBER_IN_USE
              - IDENTITY_ALREADY_MAPPED
              - PROVIDER_INVALID_RESPONSE
              - PROVIDER_RATE_LIMITED
              - USER_PROVISIONING_COMPENSATION_FAILED
              - SUPER_ADMIN_REQUIRED
              - INSUFFICIENT_ADMIN_TIER
              - CANNOT_CHANGE_OWN_ROLES
              - LAST_SUPER_ADMIN_REQUIRED
              - RESTAURANT_NOT_FOUND
              - RESTAURANT_IMAGE_NOT_FOUND
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: true
            requestId:
              type: string
...